Refactored the EmailNotificacion controller and handler structure to comply with clean code rules.
