+++ findit-backend/Handlers/EmailNotificationHandler/IEmailNotificationHandler.cs
@@ -1,0 +1,8 @@
+﻿namespace findit_backend.Handlers.EmailNotificationHandler
+{
+    public interface IEmailNotificationHandler
+    {
+        Task SendEmailAsync(string to, string subject, string body);
+
+    }
+}
