+++ findit-backend/Program.cs
@@ -1,3 +1,4 @@
+using findit_backend.Handlers.EmailNotificationHandler;
 using findit_backend.Handlers.ProductHandler;
 using findit_backend.Helpers.ProductValidationHelper;
 using findit_backend.Managers.AddToCartManager;
@@ -24,6 +25,7 @@ builder.Services.AddControllers();
 builder.Services.AddScoped<IAddToCartManager, AddToCartManager>();
 builder.Services.AddScoped<IProductValidationHelper, ProductValidationHelper>();
 builder.Services.AddScoped<IProductHandler, ProductHandler>();
+builder.Services.AddScoped<IEmailNotificationHandler, EmailNotificationHandler>();
 
 // Learn more about configuring Swagger/OpenAPI at https://aka.ms/aspnetcore/swashbuckle
 builder.Services.AddEndpointsApiExplorer();
