+++ findit-backend/appsettings.json
@@ -7,7 +7,7 @@
     },
     "AllowedHosts": "*",
   "ConnectionStrings": {
-    "FindItDatabase": "Data Source=dannypc\\sqlexpress;Initial Catalog=FindIt;Integrated Security=True;Encrypt=False"
+    "FindItDatabase": "Data Source=SEBASBOSE-WIN;Initial Catalog=FindIt;Integrated Security=True;Encrypt=True;TrustServerCertificate=True"
   },
     "Jwt": {
         "Issuer": "FindIt-Backend",
